+++ examples/webbilling-demo/src/tests/tax-calculation.test.ts
@@ -15,10 +15,13 @@ import {
   TEXAS_TAX_RESPONSE,
   NEW_YORK_TAX_RESPONSE,
   ITALY_TAX_RESPONSE,
+  STRIPE_TAX_NOT_ACTIVE_RESPONSE,
+  INVALID_TAX_ORIGIN_RESPONSE,
+  MISSING_STRIPE_PERMISSION_RESPONSE,
 } from "./helpers/fixtures";
 import {
   integrationTest,
-  SKIP_TAX_CALCULATION_REAL_TESTS,
+  SKIP_TAX_REAL_TESTS,
 } from "./helpers/integration-test";
 import {
   clickPayButton,
@@ -69,9 +72,14 @@ const mockTaxCalculationRequest = async (
     `Tax calculation (${mockMode ? "mocked" : "real"})`,
     () => {
       integrationTest.skip(
-        !mockMode && SKIP_TAX_CALCULATION_REAL_TESTS,
+        !mockMode && SKIP_TAX_REAL_TESTS,
         `Tax calculation ${mockMode ? "mocked" : "real"} tests are disabled.
-        To enable, set VITE_SKIP_TAX_CALCULATION_REAL_TESTS=true in the environment variables.`,
+        To enable, set VITE_SKIP_TAX_REAL_TESTS_UNTIL=2025-02-21 in the environment variables.`,
+      );
+
+      integrationTest.skip(
+        ({ browserName }) => !mockMode && browserName !== "chromium",
+        "Real tax calculation tests only run in Chromium",
       );
 
       integrationTest.beforeEach(async ({ page }) => {
@@ -79,6 +87,16 @@ const mockTaxCalculationRequest = async (
           // Prevent the real requests from being performed
           await page.route(TAX_ROUTE_PATH, async (route) => {
             route.abort();
+          });
+        } else {
+          // Fail the test if the rate limit is reached
+          await page.route(TAX_ROUTE_PATH, async (route) => {
+            const response = await route.fetch();
+            const json = await response.json();
+            if (json["failed_reason"] === "rate_limit_exceeded") {
+              throw new Error("Stripe Tax Calculation API rate limit reached.");
+            }
+            route.fulfill({ response, json });
           });
         }
       });
@@ -577,36 +595,20 @@ const mockTaxCalculationRequest = async (
 integrationTest.describe("Tax calculation setup errors", () => {
   integrationTest.fixme("Stripe tax not active", async ({ page, userId }) => {
     await page.route(TAX_ROUTE_PATH, async (route) => {
-      await route.fulfill({
-        json: {
-          mocked: true,
-          code: 7898,
-          message:
-            "Stripe account setup error: Stripe Tax must be active to calculate taxes.",
-        },
-        status: 422,
-      });
-    });
+      await route.fulfill(STRIPE_TAX_NOT_ACTIVE_RESPONSE);
 
-    page = await navigateToTaxesLandingUrl(page, userId);
-    const packageCards = await getPackageCards(page);
-    await startPurchaseFlow(packageCards[0]);
-    await confirmPaymentError(page, "Stripe Tax not active");
+      page = await navigateToTaxesLandingUrl(page, userId);
+      const packageCards = await getPackageCards(page);
+      await startPurchaseFlow(packageCards[0]);
+      await confirmPaymentError(page, "Stripe Tax not active");
+    });
   });
 
   integrationTest.fixme(
     "Invalid tax origin address",
     async ({ page, userId }) => {
       await page.route(TAX_ROUTE_PATH, async (route) => {
-        await route.fulfill({
-          json: {
-            mocked: true,
-            code: 7899,
-            message:
-              "Stripe account setup error: Origin address for Stripe Tax is missing or invalid.",
-          },
-          status: 422,
-        });
+        await route.fulfill(INVALID_TAX_ORIGIN_RESPONSE);
       });
 
       page = await navigateToTaxesLandingUrl(page, userId);
@@ -620,15 +622,7 @@ integrationTest.describe("Tax calculation setup errors", () => {
     "Missing Stripe permission",
     async ({ page, userId }) => {
       await page.route(TAX_ROUTE_PATH, async (route) => {
-        await route.fulfill({
-          json: {
-            mocked: true,
-            code: 7900,
-            message:
-              "Stripe account setup error: Required permission is missing.",
-          },
-          status: 422,
-        });
+        await route.fulfill(MISSING_STRIPE_PERMISSION_RESPONSE);
       });
 
       page = await navigateToTaxesLandingUrl(page, userId);