+++ src/stories/pages/success-page.stories.svelte
@@ -61,6 +61,7 @@
     isInElement={context.globals.viewport === "embedded"}
     onError={() => {}}
     onClose={() => {}}
+    managementUrl="http://test.com"
   />
 {/snippet}